geodjango_ca1/requirements.txt
asgiref==3.3.0
Django==3.1.3
django-crispy-forms==1.9.2
django-environ==0.4.5
django-leaflet==0.27.1
django-pwa==1.0.10
psycopg2-binary==2.8.6
python-google-places==1.4.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
uWSGI==2.0.19.1
whitenoise==5.2.0
